+++ src/main/java/cl/praxis/inmobiliaria/restController/TipoPropiedadRestController.java
@@ -2,8 +2,9 @@ package cl.praxis.inmobiliaria.restController;
 
 import cl.praxis.inmobiliaria.entities.TipoPropiedad;
 import cl.praxis.inmobiliaria.services.ICrudGenericoService;
-import cl.praxis.inmobiliaria.services.impl.TipoPropiedadService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -20,8 +21,13 @@ public class TipoPropiedadRestController {
     }
 
     @GetMapping("/tipo")
-    public TipoPropiedad obtenerTipoPropiedad(@RequestParam int id) {
-        return tipoPropiedadService.buscarPorId(id);
+    public ResponseEntity<?> obtenerTipoPropiedad(@RequestParam int id) {
+        TipoPropiedad tipoEncontrado = tipoPropiedadService.buscarPorId(id);
+        if (tipoEncontrado == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Tipo Propiedad no encontrado");
+        }
+        return ResponseEntity.ok(tipoEncontrado);
     }
 
     @PostMapping("/nuevo")
@@ -30,12 +36,24 @@ public class TipoPropiedadRestController {
     }
 
     @PutMapping("/actualizar")
-    public TipoPropiedad actualizarTipoPropiedad(@RequestBody TipoPropiedad tipoPropiedad) {
-        return tipoPropiedadService.actualizar(tipoPropiedad);
+    public ResponseEntity<?> actualizarTipoPropiedad(@RequestBody TipoPropiedad tipoPropiedad) {
+        TipoPropiedad tipoEncontrado = tipoPropiedadService.buscarPorId(tipoPropiedad.getId());
+        if(tipoEncontrado == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Tipo Propiedad no encontrado");
+        }
+        tipoPropiedadService.actualizar(tipoPropiedad);
+        return ResponseEntity.ok(tipoEncontrado);
     }
 
     @DeleteMapping("/eliminar")
-    public void eliminarTipoPropiedad(@RequestBody TipoPropiedad tipoPropiedad) {
+    public ResponseEntity<?> eliminarTipoPropiedad(@RequestBody TipoPropiedad tipoPropiedad) {
+        TipoPropiedad tipoEncontrado = tipoPropiedadService.buscarPorId(tipoPropiedad.getId());
+        if(tipoEncontrado == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Tipo Propiedad no encontrado");
+        }
         tipoPropiedadService.remover(tipoPropiedad);
+        return ResponseEntity.ok(tipoEncontrado);
     }
 }
